requirements.txt
# Intelli-Credit Dependencies
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Document Processing
python-docx>=1.1.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Web Scraping & NLP
beautifulsoup4>=4.12.0
requests>=2.31.0
langchain>=0.1.0
langchain-community>=0.0.10

# Graph Analytics
networkx>=3.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.43.0

# LLM Integration
anthropic>=0.18.0
openai>=1.0.0

# API & Backend
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
